requirements.txt
fastapi
uvicorn
pillow
beautifulsoup4
requests
python-dotenv
chromadb
torch
transformers
python-multipart
markdown